requirements.txt
# Python 3.12-compatible requirements - EXACT versions
numpy==1.26.4
pandas==2.2.1
streamlit==1.32.0
plotly==5.19.0
fpdf2==2.7.8
python-dateutil==2.8.2
joblib==1.3.2
scikit-learn==1.4.1.post1
scipy==1.12.0
xgboost==2.0.3
python-pptx==1.0.1
openpyxl==3.1.2